+++ peter_shirley/color.h
@@ -5,7 +5,7 @@
 typedef union u_rgb		t_rgb;
 
 t_rgb	rgb(uint8_t r, uint8_t g, uint8_t b, uint8_t alpha);
-t_rgb	rgb_mul(t_rgb col, float alpha);
+t_rgb	rgb_mul(t_rgb col, double alpha);
 t_rgb	rgb_add(t_rgb col1, t_rgb col2);
 
 t_rgb	red(void);
